+++ src/web/src/modules/contact/router/index.js
@@ -1,0 +1,15 @@
+const routes = [
+  {
+    path: "",
+    component: () => import("@/layouts/Default.vue"),
+    children: [
+      {
+        path: "/contact",
+        name: "contact",
+        component: () => import("../views/Contact.vue"),
+      },
+    ],
+  },
+];
+
+export default routes;